- Update switch digital input assignments as soldered/connected at Brian's house. - Configure encoder interrupt at startup. - Refactor menus to use enums instead of hardcoded numbers. - Add testMode screens for raw battery readings and drive motors. - Fix battery voltage functions (needed an explicit cast to u32). - Update encoder thresholds. - Add printCharN() function to LCD driver. - Fix s32 typedef defined incorrectly. - Increase buttonHeld() delay to 400ms.

+++ trunk/Code/PolyBotLibrary/globals.h
@@ -34,7 +34,7 @@ typedef int8_t   s08; //!< Signed 8-bit integer, range: -128 to +127
 typedef uint16_t u16; //!< Unsigned 16-bit integer, range: 0 to +65,535
 typedef int16_t  s16; //!< Signed 16-bit integer, range: -32,768 to +32,767
 typedef uint32_t u32; //!< Unsigned 32-bit integer, range: 0 to +4,294,967,295
-typedef int16_t  s32; //!< Signed 32-bit integer, range: -2,147,483,648 to +2,147,483,647
+typedef int32_t  s32; //!< Signed 32-bit integer, range: -2,147,483,648 to +2,147,483,647
 
 //Bit manipulation macros
 #define sbi(a, b) ((a) |= 1 << (b))       //!< Sets bit b in variable a.
